Main: Rect - use int32 as base type for MSVC/ GCC consistency

+++ OgreMain/include/OgreCommon.h
@@ -715,7 +715,7 @@ namespace Ogre {
 
         /** Structure used to define a rectangle in a 2-D integer space.
         */
-        typedef TRect< long > Rect;
+        typedef TRect< int32 > Rect;
 
         /** Structure used to define a box in a 3-D integer space.
             Note that the left, top, and front edges are included but the right,
